move bfabric_config.py into config/config_file.py
